Surface Selected Rows on TableBody
Since the selected attribute on the rows has absolutely no effect on actual selected rows, when it changes the render is never actually updated. With this surfaced the developer can supply the array of indexes that we are already checking index in for each seperate row, and it will automatically update the table the way you would expect.